Gui[NotWorking]/requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==40.0.2
dnspython==2.3.0
evdev==1.6.1
idna==3.4
#maskpass==0.3.7
#numpy==1.22
#paramiko==2.8.1
plyer==2.1.0
psutil==5.9.2
pycparser==2.21
PyNaCl==1.5.0
pynput==1.7.6
#pyshorteners==1.0.1
PySide6==6.4.2
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
python-xlib==0.33
PyYAML==6.0
requests==2.27.1
scapy==2.4.4
six==1.16.0
#speedtest==0.0.1
#speedtest-cli==2.1.3
urllib3==1.26.15
pyqtdarktheme==2.1.0
nuitka==1.5.7
fastjsonschema==2.17.1
Flask==2.3.2